+++ src/main/java/simulation/strategy/ConservativeStrategy.java
@@ -4,7 +4,7 @@ package simulation.strategy;
  * The least random type of simulation.
  */
 public class ConservativeStrategy implements StrategyType {
-  private static final int   foodConsumption = 10;
+  private static final float   foodConsumption = 10.0f;
   private static final float priceMultiplier = 1.0f;
   private static final float travelCost      = 1.15f;
 
@@ -13,7 +13,7 @@ public class ConservativeStrategy implements StrategyType {
    * @return daily food consumption in units of food.
    */
   @Override
-  public int getFoodConsumption() {
+  public float getFoodConsumption() {
     return foodConsumption;
   }
